+++ src/index.ts
@@ -96,7 +96,7 @@ nextui
   .option('-p --packagePath [string]', 'Specify the path to the package.json file')
   .option('-tw --tailwindPath [string]', 'Specify the path to the tailwind.config.js file')
   .option('-app --appPath [string]', 'Specify the path to the App.tsx file')
-  .option('--prettier [boolean]', 'Apply Prettier formatting to the added content', false)
+  .option('--prettier [boolean]', 'Apply Prettier formatting to the added content')
   .option('--addApp [boolean]', 'Include App.tsx file content that requires a provider', false)
   .action(addAction);
 
@@ -120,6 +120,7 @@ nextui
   .option('-p --packagePath [string]', 'Specify the path to the package.json file')
   .option('-a --all [boolean]', 'Remove all components', false)
   .option('-tw --tailwindPath [string]', 'Specify the path to the tailwind.config.js file')
+  .option('--prettier [boolean]', 'Apply Prettier formatting to the added content')
   .action(removeAction);
 
 nextui