+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [0.3.2](https://github.com/nextui-org/nextui-cli/compare/v0.3.1...v0.3.2) (2024-06-14)
+
+### Bug Fixes
+
+- add command error in tailwind ([#80](https://github.com/nextui-org/nextui-cli/issues/80)) ([455d6e0](https://github.com/nextui-org/nextui-cli/commit/455d6e03ed33f2bba18cb5cd113966af0a7dc69f))
+- upgrade command select option cannot select @nextui-org/react pkg ([#78](https://github.com/nextui-org/nextui-cli/issues/78)) ([93e5cdc](https://github.com/nextui-org/nextui-cli/commit/93e5cdc98299ec75d5720c0a6d44d08241c66dc7))
+
+### Features
+
+- init command add invalid options check ([#79](https://github.com/nextui-org/nextui-cli/issues/79)) ([1573e48](https://github.com/nextui-org/nextui-cli/commit/1573e481f15d8bb0660b0037c226799898285d86))
+- optimize upgrade select options package name display ([#76](https://github.com/nextui-org/nextui-cli/issues/76)) ([5c3c3f7](https://github.com/nextui-org/nextui-cli/commit/5c3c3f70bb75b80157075e4817a9ea47c67fa573))
+- upgrade command add setup pnpm ([#77](https://github.com/nextui-org/nextui-cli/issues/77)) ([1beb52c](https://github.com/nextui-org/nextui-cli/commit/1beb52ce1b88583649aa7ae837780d708eab31eb))
+
 ## [0.3.1](https://github.com/nextui-org/nextui-cli/compare/v0.3.0...v0.3.1) (2024-06-01)
 
 ### Bug Fixes